Update default recrding type to float32

+++ src/recorder/main.py
@@ -20,7 +20,7 @@ def __change_dir_to_pkg_root() -> None:
 
 def record(duration: int, frequency: int = 44100, channels: int = 2):
     recording = cast(
-        numpy.ndarray[Any, numpy.dtype[numpy.float64]],
+        numpy.ndarray[Any, numpy.dtype[numpy.float32]],
         sounddevice.rec(
             int(duration * frequency), samplerate=frequency, channels=channels
         ),
